refactored inventory and items api
