+++ MTM-Backend/src/routes/v1/inventory/inventory.service.ts
@@ -1,5 +1,8 @@
 import { db } from "../../../utils/db.server";
-import type { InventoryType } from "../../../types/inventory";
+import type {
+  InventoryType,
+  InventoryInputType,
+} from "../../../types/inventory";
 
 /**
  * Get inventory by page
@@ -15,4 +18,34 @@ export const getInventoryByPage = async (
     skip: (page - 1) * pageSize,
     take: pageSize,
   });
+};
+
+/**
+ * Create a item in the inventory table
+ * @param item
+ * @returns the created item
+ */
+export const createItem = async (
+  item: InventoryInputType,
+): Promise<InventoryType> => {
+  const createdItem: InventoryType = await db.item.create({
+    data: {
+      category: item.category,
+      name: item.name,
+      quantityUsed: item.quantityUsed,
+      quantityNew: item.quantityNew,
+      valueNew: item.valueNew,
+      valueUsed: item.valueUsed,
+    },
+    select: {
+      id: true,
+      category: true,
+      name: true,
+      quantityUsed: true,
+      quantityNew: true,
+      valueNew: true,
+      valueUsed: true,
+    },
+  });
+  return createdItem;
 };
