+++ MTM-Backend/src/routes/v1/item/item.router.ts
@@ -3,9 +3,8 @@ import express, {
   type Response,
   type NextFunction,
 } from "express";
-import type { ItemInputNoID, ItemType } from "../../../types/item";
+import type { ItemType } from "../../../types/item";
 import * as ItemsService from "./item.service";
-import Joi from "joi";
 
 const itemsRouter = express.Router();
 
@@ -38,27 +37,6 @@ itemsRouter.get(
         items = await ItemsService.getItemsCategoryName(category, name);
       }
       return res.status(200).json(items);
-    } catch (e) {
-      next(e);
-    }
-  },
-);
-
-itemsRouter.post(
-  "/v1",
-  async (req: Request, res: Response, next: NextFunction) => {
-    const schema = Joi.object({
-      category: Joi.string().required(),
-      name: Joi.string().required(),
-      quantityUsed: Joi.number().required(),
-      quantityNew: Joi.number().required(),
-      valueNew: Joi.number().required(),
-      valueUsed: Joi.number().required(),
-    });
-    try {
-      const data = (await schema.validateAsync(req.body)) as ItemInputNoID;
-      const item = await ItemsService.createItem(data);
-      return res.status(201).json(item);
     } catch (e) {
       next(e);
     }
