+++ MTM-Backend/src/types/item.ts
@@ -3,15 +3,6 @@ export interface ItemInput {
   category: string;
 }
 
-export interface ItemInputNoID {
-  category: string;
-  name: string;
-  quantityUsed: number;
-  quantityNew: number;
-  valueNew: number;
-  valueUsed: number;
-}
-
 export interface ItemType {
   id: number;
   category: string;
